Removed unnecessary isOp check when using Bukkit Permissions. Explained:
BukkitPermissions has four values that you can set to each permission
node - true, false, op and not op. In plugin.yml I have set the default
value to use the reload command to op. This means that Bukkit's
Permissions plugin will set the value to true if the player is an op,
therefore eliminating the need to check if the player is an op if bukkit
permissions is being used.